refactored some more UI code
