+++ server/static/index.js
@@ -4,6 +4,14 @@ import {
   renderRecordingControls,
   renderArrows,
 } from './components/recorder.js';
+import {
+  disableSamplePlayer,
+  enableSamplePlayer,
+  disableDownloadLink,
+  enableDownloadLink,
+  initializeRecord,
+  updateRecordButton,
+} from './components/app.js';
 import { renderNoiseList } from './components/list.js';
 import {
   NEED_PERMISSIONS,
@@ -18,49 +26,36 @@ import {
   UI
 */
 
-// TODO: merge enable/disable fns
-function disableSamplePlayer() {
-  const player = document.querySelector('[data-id=player]');
-  player.setAttribute('disabled', 'disabled');
-  player.src = null;
+function renderRecorderAndArrows() {
+  renderRecordingControls(state.recorder);
+  renderArrows(
+    state.recorder.status,
+    state.noiseList,
+    state.selectedNoise,
+    decrementSelectedNoise,
+    incrementSelectedNoise,
+  );
 }
 
-function enableSamplePlayer({ url }) {
-  // hook up player
-  const player = document.querySelector('[data-id=player]');
-
-  player.src = url;
-  player.removeAttribute('disabled');
-}
-
-function disableDownloadLink() {
-  const downloadLink = document.querySelector('[data-id=download]');
-  downloadLink.classList.add('DownloadLink--disabled');
-  downloadLink.href = null;
-}
-
-function enableDownloadLink({ url, filename }) {
-  // hook up download link
-  const downloadLink = document.querySelector('[data-id=download]');
-
-  downloadLink.href = url;
-  downloadLink.download = filename;
-  downloadLink.classList.remove('DownloadLink--disabled');
-}
-
-function initializeRecord(onFirstClick) {
-  const recordButton = document.querySelector('[data-id=recordButton]');
-  recordButton.addEventListener('click', onFirstClick);
-}
-
-function updateRecordButton(onRecordClick) {
-  const recordButton = document.querySelector('[data-id=recordButton]');
-  const recordButtonClone = recordButton.cloneNode(true);
-
-  // get rid of original event handler by replacing button element
-  // TODO: look into other ways of doing this, including using the original reference to the handler
-  recordButton.parentNode.replaceChild(recordButtonClone, recordButton);
-  recordButtonClone.addEventListener('click', onRecordClick);
+function renderApp() {
+  renderNoiseList(
+    state.noiseList,
+    selectNoise,
+    state.selectedNoise,
+    statuses,
+    render,
+    () =>
+      state.recorder.status === RECORDING ||
+      state.recorder.status === UPLOADING,
+  );
+  renderRecorder(
+    state.noiseList[state.selectedNoise],
+    state.recorder,
+    state.noiseList,
+    state.selectedNoise,
+    decrementSelectedNoise,
+    incrementSelectedNoise,
+  );
 }
 
 /*
@@ -228,7 +223,7 @@ const handleGetUserMediaSuccess = function(stream) {
   }
 
   recordClickHandler(); // manually trigger first time we've successfully gotten permissions to the mic since the user already clicked the Record button
-  
+
   /* UI */
   updateRecordButton(recordClickHandler);
 
@@ -247,14 +242,7 @@ const handleGetUserMediaSuccess = function(stream) {
       });
 
       /* UI */
-      renderRecordingControls(state.recorder);
-      renderArrows(
-        state.recorder.status,
-        state.noiseList,
-        state.selectedNoise,
-        decrementSelectedNoise,
-        incrementSelectedNoise,
-      );
+      renderRecorderAndArrows();
     }
   });
 
@@ -269,14 +257,7 @@ const handleGetUserMediaSuccess = function(stream) {
     });
 
     /* UI */
-    renderRecordingControls(state.recorder);
-    renderArrows(
-      state.recorder.status,
-      state.noiseList,
-      state.selectedNoise,
-      decrementSelectedNoise,
-      incrementSelectedNoise,
-    );
+    renderRecorderAndArrows();
     // TODO: also render list (with isDisabled, maybe just a boolean instead of function, returning false)?
   });
 
@@ -315,7 +296,7 @@ const handleGetUserMediaSuccess = function(stream) {
 
     enableDownloadLink({
       url,
-      filename, /* from state */
+      filename /* from state */,
     });
 
     enableSamplePlayer({
@@ -378,27 +359,6 @@ const handleGetUserMediaSuccess = function(stream) {
 
 function render() {
   renderApp();
-}
-
-function renderApp() {
-  renderNoiseList(
-    state.noiseList,
-    selectNoise,
-    state.selectedNoise,
-    statuses,
-    render,
-    () =>
-      state.recorder.status === RECORDING ||
-      state.recorder.status === UPLOADING,
-  );
-  renderRecorder(
-    state.noiseList[state.selectedNoise],
-    state.recorder,
-    state.noiseList,
-    state.selectedNoise,
-    decrementSelectedNoise,
-    incrementSelectedNoise,
-  );
 }
 
 function processNoises(noises) {
